Refactor helper so length of list can be increased

+++ AlgorythmExcercises/Helpers/helper.cs
@@ -5,12 +5,12 @@ namespace Helpers
 {
     public static class helper
     {
-        public static List<int> Generate_List_of_integers(int length)
+        public static List<int> Generate_List_of_integers(double length)
         {
             Random rand = new Random();
             List<int> result = new List<int>();
 
-            for (int i = 0; i < length; i++)
+            for (double i = 0; i < length; i++)
             {
                 result.Add(rand.Next());
             }
